+++ src/test/java/com/bbangle/bbangle/store/service/StoreServiceTest.java
@@ -1,7 +1,6 @@
 package com.bbangle.bbangle.store.service;
 
 import static java.util.Collections.emptyMap;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.bbangle.bbangle.AbstractIntegrationTest;
@@ -14,10 +13,8 @@ import com.bbangle.bbangle.fixture.BoardStatisticFixture;
 import com.bbangle.bbangle.fixture.MemberFixture;
 import com.bbangle.bbangle.fixture.StoreFixture;
 import com.bbangle.bbangle.member.domain.Member;
-import com.bbangle.bbangle.page.StoreCustomPage;
 import com.bbangle.bbangle.page.StoreDetailCustomPage;
 import com.bbangle.bbangle.store.domain.Store;
-import com.bbangle.bbangle.store.dto.StoreResponseDto;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.List;
@@ -57,84 +54,6 @@ class StoreServiceTest extends AbstractIntegrationTest {
                 storeRepository.save(store);
             }
         }
-
-        @Test
-        @DisplayName("정상적으로 첫 페이지를 조회한다")
-        void getFirstPage() {
-            //given, when
-            StoreCustomPage<List<StoreResponseDto>> list = storeService.getList(
-                NULL_CURSOR,
-                NULL_MEMBER_ID
-            );
-            List<StoreResponseDto> content = list.getContent();
-            Boolean hasNext = list.getHasNext();
-
-            //then
-            assertThat(content).hasSize(20);
-            assertThat(hasNext).isTrue();
-        }
-
-        @Test
-        @DisplayName("정상적으로 마지막 페이지를 조회한다")
-        void getLastPage() {
-            //given
-            StoreCustomPage<List<StoreResponseDto>> firstPage = storeService.getList(NULL_CURSOR,
-                NULL_MEMBER_ID);
-            Long nextCursor = firstPage.getNextCursor();
-
-            //when
-            StoreCustomPage<List<StoreResponseDto>> lastPage = storeService.getList(nextCursor,
-                NULL_MEMBER_ID);
-
-            List<StoreResponseDto> lastPageContent = lastPage.getContent();
-            Boolean lastPageHasNext = lastPage.getHasNext();
-            Long lastPageNextCursor = lastPage.getNextCursor();
-
-            //then
-            assertThat(lastPageContent).hasSize(10);
-            assertThat(lastPageHasNext).isFalse();
-        }
-
-        @Test
-        @DisplayName("마지막 자료를 조회하는 경우 nextCursor는 -1을 가리킨다")
-        void getLastContent() {
-            //given, when
-            StoreCustomPage<List<StoreResponseDto>> firstPage = storeService.getList(NULL_CURSOR,
-                NULL_MEMBER_ID);
-            Long nextCursor = firstPage.getNextCursor();
-            StoreCustomPage<List<StoreResponseDto>> lastPage = storeService.getList(nextCursor,
-                NULL_MEMBER_ID);
-            Long lastContentCursor = lastPage.getNextCursor();
-
-            StoreCustomPage<List<StoreResponseDto>> noContent = storeService.getList(
-                lastContentCursor,
-                NULL_MEMBER_ID);
-
-            //then
-            assertThat(noContent.getContent()).isEmpty();
-            assertThat(noContent.getNextCursor()).isEqualTo(-1L);
-        }
-
-        @Test
-        @DisplayName("좋아요를 누른 store는 isWished가 true로 반환된다")
-        void getWishedContent() throws Exception {
-            //given, when
-            StoreCustomPage<List<StoreResponseDto>> before = storeService.getList(NULL_CURSOR,
-                NULL_MEMBER_ID);
-            StoreResponseDto first = before.getContent()
-                .stream()
-                .findFirst()
-                .orElseThrow(Exception::new);
-            wishListStoreService.save(memberId, first.getStoreId());
-
-            //then
-            StoreResponseDto wishedContent = storeService.getList(NULL_CURSOR, memberId)
-                .getContent()
-                .stream()
-                .findFirst()
-                .orElseThrow(Exception::new);
-            assertThat(wishedContent.getIsWished()).isTrue();
-        }
     }
 
     @Nested
@@ -159,7 +78,7 @@ class StoreServiceTest extends AbstractIntegrationTest {
                 "veganTag", false,
                 "ketogenicTag", false,
                 "orderStartDate", LocalDateTime.now(),
-                "soldout",true
+                "soldout", true
             ));
 
             Map<String, Object> params = new HashMap<>();
@@ -301,8 +220,10 @@ class StoreServiceTest extends AbstractIntegrationTest {
             StoreDetailCustomPage<List<BoardInfoDto>> boardsInStoreDtos = storeService.getBoardsInStore(
                 NULL_MEMBER_ID, store.getId(), NULL_CURSOR);  // board id desc 임
 
-            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(4).getIsNotification()).isTrue();
-            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(3).getIsNotification()).isFalse();
+            AssertionsForClassTypes.assertThat(
+                boardsInStoreDtos.getContent().get(4).getIsNotification()).isTrue();
+            AssertionsForClassTypes.assertThat(
+                boardsInStoreDtos.getContent().get(3).getIsNotification()).isFalse();
         }
 
         @Test
@@ -311,8 +232,10 @@ class StoreServiceTest extends AbstractIntegrationTest {
             StoreDetailCustomPage<List<BoardInfoDto>> boardsInStoreDtos = storeService.getBoardsInStore(
                 NULL_MEMBER_ID, store.getId(), NULL_CURSOR);  // board id desc 임
 
-            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(4).getIsSoldOut()).isTrue();
-            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(3).getIsSoldOut()).isFalse();
+            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(4).getIsSoldOut())
+                .isTrue();
+            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(3).getIsSoldOut())
+                .isFalse();
         }
 
         @Test
@@ -321,11 +244,11 @@ class StoreServiceTest extends AbstractIntegrationTest {
             StoreDetailCustomPage<List<BoardInfoDto>> boardsInStoreDtos = storeService.getBoardsInStore(
                 NULL_MEMBER_ID, store.getId(), NULL_CURSOR);  // board id desc 임
 
-            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(1).getIsBundled()).isTrue();
-            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(0).getIsBundled()).isFalse();
+            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(1).getIsBundled())
+                .isTrue();
+            AssertionsForClassTypes.assertThat(boardsInStoreDtos.getContent().get(0).getIsBundled())
+                .isFalse();
         }
 
     }
-
-
 }
